super.c: Separate module for SuperNOVAS only functions not needed by novas.c
